[TF FE] Fix compilation warnings

+++ src/frontends/tensorflow_common/src/pass/transpose_sinking.cpp
@@ -173,8 +173,8 @@ static void convert_binary_to_default_order(const shared_ptr<Node>& binary,
     // if right input is being implicitly broadcasted, insert a reshape
     // instead of a transpose
     shared_ptr<Node> new_node;
-    auto left_rank = get_static_rank(left);
-    if (left_rank < static_cast<int64_t>(perm_to_def.size()) && left.get_partial_shape().is_static()) {
+    auto left_rank = static_cast<size_t>(get_static_rank(left));
+    if (left_rank < perm_to_def.size() && left.get_partial_shape().is_static()) {
         auto left_shape = left.get_shape();
         left_shape.insert(left_shape.begin(), perm_to_def.size() - left_rank, 1);
